Fix warnings, from Jan Tache in GitHub issue 2692.

+++ server.c
@@ -161,7 +161,8 @@ server_tidy_event(__unused int fd, __unused short events, __unused void *data)
 
     format_tidy_jobs();
 
-    log_debug("%s: took %llu milliseconds", __func__, get_timer() - t);
+    log_debug("%s: took %llu milliseconds", __func__,
+        (unsigned long long)(get_timer() - t));
     evtimer_add(&server_ev_tidy, &tv);
 }
